requirements.txt
pandas 

scikit-learn 
flask 
streamlit 
matplotlib 
seaborn 
requests
reportlab
plotly
xgboost
textblob
gunicorn
python-dotenv 
langchain
langchain_community
openai
flask-cors 
transformers 
pdfplumber 
torch
mysql-connector-python==8.0.16   
face-recognition 
cmake

opencv-python 
numpy==1.26.4     
wheel 
setuptool
pillow 

folium 
streamlit-folium
sqlalchemy 
pymysql
bs4
debugpy 
streamlit-geolocation
selenium